completed index media queries
