+++ src/Components/HeaderIndex.js
@@ -6,7 +6,6 @@ import SearchBar from "./SearchBar";
 const HeaderIndex = () => {
   return (
     <header className="index-header">
-      {/* <MobileNav /> */}
       <Nav />
       <div className="index-header-box wrapper">
         <h2>Find & Hire Experts for any job.</h2>
@@ -29,6 +28,7 @@ const HeaderIndex = () => {
           <p>Dollar Payout</p>
         </div>
       </div>
+
     </header>
   )
 }
