+++ src/sass/partials/components/_footer.scss
@@ -32,6 +32,9 @@ footer {
   .jobi-h4 {
     font-size: $fs-xl;
   }
+        h4 {
+          cursor: pointer;
+        }
 
   a, p {
     font-size: 1.7rem;
@@ -96,6 +99,8 @@ footer {
   justify-content: space-between;
   align-items: center;
   padding: 0 0 3% 0;
+
+
   .left-links {
     flex-basis: 30%;
     a:hover,
@@ -106,7 +111,7 @@ footer {
       margin-right: 1%;
     }
   }
-  div.copyright-text {
+  .copyright-text {
     flex-basis: 40%;
     text-align: center;
     p {
