+++ src/sass/partials/components/_header.scss
@@ -12,6 +12,11 @@
   background-size: cover;
 }
 
+.header-background {
+  // min-height: 100vh;
+  background: linear-gradient(90deg, rgba(0, 0, 0, 0.6) 38%, rgba(0, 0, 0, 0.6) 100%);
+}
+
 .index-header-box {
   display: flex;
   flex-direction: column;
@@ -28,6 +33,8 @@
   p {
     color: $white;
     margin-bottom: 5%;
+    text-align: center;
+    font-weight: bold;
   }
 }
 
@@ -48,6 +55,7 @@
   justify-content: center;
   align-items: center;
   margin: 8%;
+  text-align: center;
 
   p {
     color: $white;
